tratativas para req dados apenas do usuario logado

+++ routes/transactions.py
@@ -1,6 +1,7 @@
 from flask import request, jsonify, Blueprint
-from flask_login import login_required
-from models import Transaction
+from flask_login import current_user, login_required
+from flask_migrate import current
+from models import Transaction, User
 from extensions import db
 
 transactions_bp = Blueprint('transactions',__name__, url_prefix='/api/transactions')
@@ -10,7 +11,7 @@ transactions_bp = Blueprint('transactions',__name__, url_prefix='/api/transactio
 def add_transactions():
   data = request.json
 
-  if data.get('description') and data.get('value') and data.get('type') and data.get('date'):
+  if data.get('description') and data.get('amount') and data.get('type') and data.get('date'):
 
     if data['type'] != 'receita' and data['type'] != 'despesa':
 
@@ -21,6 +22,7 @@ def add_transactions():
       amount = data['amount'],
       type = data['type'],
       date = data['date'],
+      user_id = current_user.id
     )
 
     db.session.add(transaction)
@@ -28,14 +30,16 @@ def add_transactions():
 
     return jsonify({
       'message': 'Transaction added successfully',
+      'user': current_user.id
     })
   
   return jsonify({'message': 'Invalid transaction data'}), 400
 
 
 @transactions_bp.route('/', methods=['GET'])
+@login_required
 def transactions():
-  transactions = Transaction.query.all()
+  transactions = Transaction.query.filter_by(user_id=current_user.id)
 
   all_transactions = []
 
@@ -47,7 +51,8 @@ def transactions():
         'description': transaction.description,
         'amount': transaction.amount,
         'type': transaction.type,
-        'date': transaction.date
+        'date': transaction.date,
+        'user_id': transaction.user_id
       })
 
     return jsonify(all_transactions)
@@ -55,10 +60,11 @@ def transactions():
   return jsonify({'message': 'Transactions not found'})
 
 @transactions_bp.route('/delete/<int:transaction_id>', methods=['DELETE'])
+@login_required
 def delete_transaction(transaction_id):
   transaction = Transaction.query.get(transaction_id)
 
-  if transaction:
+  if transaction and transaction.user_id == current_user.id:
     db.session.delete(transaction)
     db.session.commit()
 
@@ -67,12 +73,13 @@ def delete_transaction(transaction_id):
   return jsonify({'message': 'Transaction not found'}), 404
 
 @transactions_bp.route('/update/<int:transaction_id>', methods=['PUT'])
+@login_required
 def update_transaction(transaction_id):
   data = request.json
 
   transaction = Transaction.query.get(transaction_id)
 
-  if transaction:
+  if transaction and transaction.user_id == current_user.id:
     transaction.description = data.get('description', transaction.description)
     transaction.amount = data.get('amount', transaction.amount)
     transaction.type = data.get('type', transaction.type)
